% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prms_data.R
\name{read_prms_data}
\alias{read_prms_data}
\title{Read in a PRMS data file}
\usage{
read_prms_data(path = paste(getwd(), "input/input.data", sep = "/"),
  col_names = NULL, NA_flag = -999)
}
\arguments{
\item{path}{path to the data file; the default searches for the data file in an "input" subfolder of the working directory}

\item{col_names}{optional character vector specifying the column names.}

\item{NA_flag}{optional flag specifying the data values that need to be converted to NA; defaults to -999}
}
\value{
a data frame
}
\description{
\code{read_prms_data} reads in a PRMS data file and returns it as a data frame
}
\references{
\emph{Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., https://dx.doi.org/10.3133/tm6B7}, \url{https://pubs.usgs.gov/tm/6b7/pdf/tm6-b7.pdf}
}
