% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prms_map.R
\name{read_prms_map}
\alias{read_prms_map}
\title{Read a PRMS Map Results output file}
\usage{
read_prms_map(path = NULL, nrow = NULL, ncol = NULL, start_date = NULL)
}
\arguments{
\item{path}{path to the output file}

\item{nrow}{number of rows in the mapped spatial resolution}

\item{ncol}{number of columns in the mapped spatial resolution}

\item{start_date}{Date object specifying the start date of the output}
}
\value{
a list with (1) a 3D array of dimensions \code{nrow x ncol x nstp} with nstp the amount of time steps, (2) vectors of start and end dates of each time step, (3) number of days in each time step and (4) means of the variable averaged over each time step
}
\description{
\code{read_prms_map} reads in a PRMS Map Results output file which maps selected PRMS variables to specified spatial and temporal distributions and returns it as an 3D array of dimensions \code{nrow x ncol x nstp}. Additionally, the start dates, end dates and number of days of each time step and the means of the variable averaged over the time step are returned as well.
}
\references{
\emph{Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., https://dx.doi.org/10.3133/tm6B7}, \url{https://pubs.usgs.gov/tm/6b7/pdf/tm6-b7.pdf}
}
