% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prms_output.R
\name{read_prms_output}
\alias{read_prms_output}
\title{Read a PRMS csv output file}
\usage{
read_prms_output(path = paste0(getwd(), "output.csv"), as_date = T)
}
\arguments{
\item{path}{path to the csv file}

\item{as_date}{logical, indicating if the first column holding the simulation date should be of class \code{Date}}
}
\value{
a data frame
}
\description{
\code{read_prms_output} reads a PRMS output file in 'comma separated value' format (.csv) and returns it as a data frame
}
\references{
\emph{Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., https://dx.doi.org/10.3133/tm6B7}, \url{https://pubs.usgs.gov/tm/6b7/pdf/tm6-b7.pdf}
}
