% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prms_parms.R
\name{read_prms_parms}
\alias{read_prms_parms}
\title{Read a PRMS parameter file}
\usage{
read_prms_parms(path = paste(getwd(), "input/input.params", sep = "/"))
}
\arguments{
\item{path}{path to the parameter file}
}
\value{
an PRMS Parameter Object which is a list with 3 elements: comments (character vector), dimensions (list with names and values) and parameters (list with names, number of dimensions, dimensions, number of values, types and values)
}
\description{
\code{read_prms_parms} reads in a PRMS parameter file and returns an PRMS Parameter Object with comments, dimensions and parameter elements
}
\references{
\emph{Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., https://dx.doi.org/10.3133/tm6B7}, \url{https://pubs.usgs.gov/tm/6b7/pdf/tm6-b7.pdf}
}
\seealso{
\code{\link{create_prms_parms}}, \code{\link{write_prms_parms}}
}
