% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prms.R
\name{run_prms}
\alias{run_prms}
\title{Run a PRMS model}
\usage{
run_prms(run_directory = paste(getwd(), "&", "cd .."),
  exe_path = "../../bin/prmsIV.exe",
  control_path = "./control/nete.control")
}
\arguments{
\item{run_directory}{optional, the full path to the project directory where the output will be stored}

\item{exe_path}{path to the PRMS executable. May be relative with respect to \code{run_directory}}

\item{control_path}{path to the control file. May be relative with respect to \code{run_directory}}
}
\description{
\code{run_prms} runs a PRMS model by opening a Windows shell and passing the control file to the PRMS executable.
}
\references{
\emph{Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., https://dx.doi.org/10.3133/tm6B7}, \url{https://pubs.usgs.gov/tm/6b7/pdf/tm6-b7.pdf}
}
