% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_prms_parms.R
\name{write_prms_parms}
\alias{write_prms_parms}
\title{Write a PRMS parameter file}
\usage{
write_prms_parms(prms_parms, path = paste(getwd(), "input/input.parms", sep =
  "/"))
}
\arguments{
\item{prms_parms}{a PRMS Parameter Object}

\item{path}{to write to}
}
\description{
\code{write_prms_parms} writes a PRMS parameter file based on a PRMS Parameter Object
}
\references{
\emph{Markstrom, S.L., Regan, R.S., Hay, L.E., Viger, R.J., Webb, R.M.T., Payn, R.A., and LaFontaine, J.H., 2015, PRMS-IV, the precipitation-runoff modeling system, version 4: U.S. Geological Survey Techniques and Methods, book 6, chap. B7, 158 p., https://dx.doi.org/10.3133/tm6B7}, \url{https://pubs.usgs.gov/tm/6b7/pdf/tm6-b7.pdf}
}
\seealso{
\code{\link{read_prms_parms}}, \code{\link{create_prms_parms}}
}
